COG-1445: iterator package is renamed to paginator
