+++ jobs-cap-users/src/main/java/gov/ca/cwds/jobs/cap/users/paginator/CapUsersInitialJobPaginator.java
@@ -1,4 +1,4 @@
-package gov.ca.cwds.jobs.cap.users.iterator;
+package gov.ca.cwds.jobs.cap.users.paginator;
 
 import com.google.inject.Inject;
 import gov.ca.cwds.idm.dto.UsersPage;
